// mainframe_platform_frontend/src/hooks/useLLMOutput/type.ts
type LLMOutputMatch = {
  startIndex: number;
  endIndex: number;
  outputRaw: string;
};
type BlockMatch = MatchBase & LLMOutputMatchWithLookBack;
type MatchBase = {
  block: LLMOutputFallbackBlock;
  priority: number;
  llmOutput: string;
  isComplete: boolean;
};
type MatchWithLLMOutput = MatchBase & { match: LLMOutputMatch };
type LLMOutputMatchWithLookBack = LLMOutputMatch & {
  output: string;
  visibleText: string;
  isVisible: boolean;
};
type MaybeLLMOutputMatch = LLMOutputMatch | undefined;
type LLMOutputComponent<Props = unknown> = React.FC<
  Props & {
    blockMatch: BlockMatch;
  }
>;
type LLMOutputMatcher = (llmOutput: string) => MaybeLLMOutputMatch;
type LookBack = {
  output: string;
  visibleText: string;
};
type LookBackFunctionParams = {
  output: string;
  isComplete: boolean;
  visibleTextLengthTarget: number;
  isStreamFinished: boolean;
};
type LookBackFunction = (params: LookBackFunctionParams) => LookBack;
type LLMOutputFallbackBlock = {
  component: LLMOutputComponent;
  lookBack: LookBackFunction;
};
type LLMOutputBlock = {
  findPartialMatch: LLMOutputMatcher;
  findCompleteMatch: LLMOutputMatcher;
} & LLMOutputFallbackBlock;
type ThrottleParams = {
  outputRaw: string;
  outputRendered: string;
  outputAll: string;
  visibleText: string;
  visibleTextAll: string;
  isStreamFinished: boolean;
  frameCount: number;
  frameTime: DOMHighResTimeStamp;
  frameTimePrevious: DOMHighResTimeStamp | undefined;
  startStreamTime: DOMHighResTimeStamp;
  finishStreamTime: DOMHighResTimeStamp | undefined;
  visibleTextLengthsAll: number[];
  outputLengths: number[];
  visibleTextIncrements: number[];
  visibleTextLengthTarget: number;
};
type ThrottleResponse = {
  visibleTextIncrement: number;
};
type ThrottleFunction = (params: ThrottleParams) => ThrottleResponse;

type LLMOutputProps = {
  llmOutput: string;
  blocks?: LLMOutputBlock[];
  fallbackBlock: LLMOutputFallbackBlock;
  isStreamFinished: boolean;
  throttle?: ThrottleFunction;
  onFinish?: () => void;
};
type UseLLMOutputReturn = {
  blockMatches: BlockMatch[];
  isFinished: boolean;
  finishCount: number;
  visibleText: string;
  restart: () => void;
};
// ...
type ThrottleBasicOptions = {
  readAheadChars: number;
  targetBufferChars: number;
  adjustPercentage: number;
  frameLookBackMs: number;
  windowLookBackMs: number;
};